requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
fonttools==4.28.5
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
msgpack==1.0.3
numpy==1.22.1
packaging==21.3
Pillow==9.0.0
py-algorand-sdk==1.9.0
pycparser==2.21
pycryptodomex==3.13.0
PyNaCl==1.5.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
telebot==0.0.4
telegram-send==0.25
tinyman-py-sdk @ git+https://github.com/tinymanorg/tinyman-py-sdk.git@c922cba801e3bdab0f3a5a5e19a25c292f16a9fd
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
